Moving mantle application to standalone package
