+++ src/mantle/application/composer.json
@@ -1,0 +1,37 @@
+{
+    "name": "mantle-framework/application",
+    "description": "The Mantle Framework Application Package",
+    "type": "project",
+    "require": {
+        "php": "^8.0",
+        "alleyinteractive/composer-wordpress-autoloader": "^1.0",
+        "mantle-framework/contracts": "^0.9",
+        "mantle-framework/support": "^0.9"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "0.9-dev"
+        },
+        "wordpress-autoloader": {
+            "autoload": {
+                "Mantle\\Application": "./"
+            }
+        }
+    },
+    "license": "GPL-2.0-or-later",
+    "authors": [
+        {
+            "name": "Alley",
+            "email": "info@alley.co"
+        }
+    ],
+    "autoload": {
+        "files": [
+            "autoload.php"
+        ]
+    },
+    "minimum-stability": "dev",
+    "config": {
+        "sort-packages": true
+    }
+}
